refactor drag and drop pages
